+++ Sprint_4/App/app/_app.js
@@ -1,0 +1,12 @@
+import { Toaster } from 'react-hot-toast'
+
+function MyApp({ Component, pageProps }) {
+  return (
+    <>
+      <Component {...pageProps} />
+      <Toaster position="top-right" />
+    </>
+  )
+}
+
+export default MyApp
