+++ Sprint_4/App/app/components/auth/Footer.js
@@ -1,0 +1,11 @@
+'use client';
+
+export default function Footer() {
+  return (
+    <footer className="bg-white/70 backdrop-blur-lg py-8">
+      <div className="container mx-auto px-6 text-center text-gray-800">
+        <p>&copy; 2024 Peer Assessment System. All rights reserved.</p>
+      </div>
+    </footer>
+  )
+}
