+++ Sprint_4/App/app/components/home/CTASection.js
@@ -1,0 +1,39 @@
+'use client'
+
+import { motion } from 'framer-motion'
+import { Button } from '@/components/ui/button'
+import Link from 'next/link'
+
+export default function CTASection() {
+  return (
+    <section className="py-20">
+      <div className="container mx-auto px-6 text-center">
+        <motion.h2
+          className="text-4xl md:text-5xl font-bold mb-8 text-white"
+          initial={{ opacity: 0, y: 20 }}
+          whileInView={{ opacity: 1, y: 0 }}
+          transition={{ duration: 0.5 }}
+          viewport={{ once: true }}
+        >
+          Ready to Transform Team Assessments?
+        </motion.h2>
+        <motion.div
+          initial={{ opacity: 0, y: 20 }}
+          whileInView={{ opacity: 1, y: 0 }}
+          transition={{ duration: 0.5, delay: 0.2 }}
+          viewport={{ once: true }}
+        >
+          <Button
+            size="lg"
+            asChild
+            className="bg-gradient-to-r from-blue-500 to-purple-600 hover:from-blue-600 hover:to-purple-700 text-white shadow-lg rounded-full"
+          >
+            <Link href="/signup">
+              Get Started Now
+            </Link>
+          </Button>
+        </motion.div>
+      </div>
+    </section>
+  )
+}
