+++ Sprint_4/App/app/components/home/Footer.js
@@ -1,0 +1,11 @@
+'use client';
+
+export default function Footer() {
+  return (
+    <footer className="bg-white/10 backdrop-blur-lg py-8">
+      <div className="container mx-auto px-6 text-center text-white">
+        <p>&copy; 2024 Peer Assessment System. All rights reserved.</p>
+      </div>
+    </footer>
+  )
+}
