+++ Sprint_4/App/app/components/home/HeroSection.js
@@ -1,0 +1,76 @@
+'use client'
+
+import { motion } from 'framer-motion'
+import { Button } from '@/components/ui/button'
+import Link from 'next/link'
+
+const title = 'Peer Assessment System'
+const letters = title.split('')
+
+const containerVariants = {
+  hidden: { opacity: 0 },
+  visible: { 
+    opacity: 1,
+    transition: {
+      staggerChildren: 0.05,
+    },
+  },
+}
+
+const letterVariants = {
+  hidden: { opacity: 0, y: 50 },
+  visible: { 
+    opacity: 1, 
+    y: 0,
+    transition: {
+      type: 'spring',
+      stiffness: 500,
+      damping: 30,
+    },
+  },
+}
+
+export default function HeroSection() {
+  return (
+    <section className="min-h-screen flex items-center justify-center">
+      <div className="container mx-auto px-6 text-center">
+        <motion.h1
+          className="text-6xl md:text-8xl font-extrabold mb-6 text-white"
+          variants={containerVariants}
+          initial="hidden"
+          animate="visible"
+        >
+          {letters.map((letter, index) => (
+            <motion.span key={index} variants={letterVariants}>
+              {letter}
+            </motion.span>
+          ))}
+        </motion.h1>
+        <motion.p
+          className="text-xl md:text-2xl mb-12 max-w-2xl mx-auto text-gray-200"
+          initial={{ opacity: 0, y: 20 }}
+          animate={{ opacity: 1, y: 0 }}
+          transition={{ duration: 0.8, delay: 0.2 }}
+        >
+          Revolutionize team dynamics with our cutting-edge platform.
+          Evaluate, collaborate, and excel together.
+        </motion.p>
+        <motion.div
+          initial={{ opacity: 0, y: 20 }}
+          animate={{ opacity: 1, y: 0 }}
+          transition={{ duration: 0.8, delay: 0.4 }}
+        >
+          <Button
+            size="lg"
+            asChild
+            className="bg-gradient-to-r from-blue-500 to-purple-600 hover:from-blue-600 hover:to-purple-700 text-white shadow-lg rounded-full"
+          >
+            <Link href="/signup">
+              Get Started
+            </Link>
+          </Button>
+        </motion.div>
+      </div>
+    </section>
+  )
+}
