+++ Sprint_4/App/app/components/home/Navbar.js
@@ -1,0 +1,27 @@
+'use client';
+
+import Link from 'next/link'
+import { Button } from '@/components/ui/button'
+
+export default function Navbar() {
+  return (
+    <nav className="fixed top-0 left-0 right-0 z-50 bg-white/10 backdrop-blur-lg">
+      <div className="container mx-auto px-6 py-3 flex justify-between items-center">
+        <Link href="/" className="text-2xl font-bold text-white">
+          PeerAssess
+        </Link>
+        <div className="space-x-4">
+          <Button variant="ghost" asChild className="text-white">
+            <Link href="/login">Login</Link>
+          </Button>
+          <Button
+            asChild
+            className="bg-gradient-to-r from-blue-500 to-purple-600 hover:from-blue-600 hover:to-purple-700 text-white shadow-lg rounded-full"
+          >
+            <Link href="/signup">Sign Up</Link>
+          </Button>
+        </div>
+      </div>
+    </nav>
+  )
+}
