+++ Sprint_4/App/app/components/home/data.js
@@ -1,0 +1,71 @@
+export const features = [
+    {
+      icon: 'Users',
+      title: 'Team Management',
+      description: 'Easily create and manage teams.',
+    },
+    {
+      icon: 'CheckCircle',
+      title: 'Peer Assessment',
+      description: 'Intuitive interface for evaluating teammates.',
+    },
+    {
+      icon: 'BarChart',
+      title: 'Analytics Dashboard',
+      description: 'Comprehensive views of assessment results.',
+    },
+    {
+      icon: 'Lock',
+      title: 'Secure & Private',
+      description: 'Your data is encrypted and protected.',
+    },
+  ]
+  
+  export const testimonials = [
+    {
+      quote: 'This platform has revolutionized how we handle team projects!',
+      author: 'Prof. Jane Smith, Computer Science',
+    },
+    {
+      quote: 'The peer assessment process is now smooth and insightful.',
+      author: 'John Doe, Student',
+    },
+    {
+      quote: "It's made managing multiple project teams so much easier.",
+      author: 'Dr. Alex Johnson, Engineering',
+    },
+  ]
+  
+  export const teamMembers = [
+    {
+      name: 'Alvin Biju',
+      role: 'Backend Developer',
+      id: '40278182',
+    },
+    {
+      name: 'Othmane Sbi',
+      role: 'Backend Developer',
+      id: '40249134',
+    },
+    {
+      name: 'Youssef Youssef',
+      role: 'Backend Developer',
+      id: '40285384',
+    },
+    {
+      name: 'William White',
+      role: 'Frontend Developer',
+      id: '40135771',
+    },
+    {
+      name: 'Samy Belmihoub',
+      role: 'Frontend Developer',
+      id: '40251504',
+    },
+    {
+      name: 'John Kaspar',
+      role: 'Frontend Developer',
+      id: '40285931',
+    },
+  ]
+
