+++ Sprint_4/App/components/ui/ConfirmationModal.js
@@ -1,0 +1,69 @@
+// components/ui/ConfirmationModal.js
+'use client'
+
+import { Dialog, Transition } from '@headlessui/react'
+import { Fragment } from 'react'
+import { ExclamationIcon } from '@heroicons/react/outline'
+import { Button } from '@/components/ui/button'
+
+export default function ConfirmationModal({ isOpen, onClose, onConfirm, title, message }) {
+  return (
+    <Transition.Root show={isOpen} as={Fragment}>
+      <Dialog as="div" className="fixed z-50 inset-0 overflow-y-auto" onClose={onClose}>
+        <div className="flex items-center justify-center min-h-screen px-4 text-center">
+          {/* Background overlay */}
+          <Transition.Child
+            as={Fragment}
+            enter="ease-out duration-300"
+            enterFrom="opacity-0"
+            enterTo="opacity-50"
+            leave="ease-in duration-200"
+            leaveFrom="opacity-50"
+            leaveTo="opacity-0"
+          >
+            <Dialog.Overlay className="fixed inset-0 bg-gray-500 transition-opacity" />
+          </Transition.Child>
+
+          {/* Modal panel */}
+          <Transition.Child
+            as={Fragment}
+            enter="ease-out duration-300"
+            enterFrom="opacity-0 translate-y-4 scale-95"
+            enterTo="opacity-100 translate-y-0 scale-100"
+            leave="ease-in duration-200"
+            leaveFrom="opacity-100 translate-y-0 scale-100"
+            leaveTo="opacity-0 translate-y-4 scale-95"
+          >
+            <div className="bg-white rounded-lg overflow-hidden shadow-xl transform transition-all max-w-lg w-full p-6">
+              <div className="sm:flex sm:items-start">
+                <div className="mx-auto flex-shrink-0 flex items-center justify-center h-12 w-12 rounded-full bg-red-100 sm:mx-0">
+                  <ExclamationIcon className="h-6 w-6 text-red-600" aria-hidden="true" />
+                </div>
+                <div className="mt-3 text-center sm:mt-0 sm:ml-4 sm:text-left">
+                  <Dialog.Title as="h3" className="text-lg leading-6 font-medium text-gray-900">
+                    {title}
+                  </Dialog.Title>
+                  <div className="mt-2">
+                    <p className="text-sm text-gray-500">{message}</p>
+                  </div>
+                </div>
+              </div>
+              <div className="mt-5 sm:mt-4 sm:flex sm:flex-row-reverse">
+                <Button onClick={onConfirm} className="w-full sm:ml-3 sm:w-auto">
+                  Confirm
+                </Button>
+                <Button
+                  onClick={onClose}
+                  variant="outline"
+                  className="mt-3 w-full sm:mt-0 sm:w-auto"
+                >
+                  Cancel
+                </Button>
+              </div>
+            </div>
+          </Transition.Child>
+        </div>
+      </Dialog>
+    </Transition.Root>
+  )
+}
