+++ Sprint_4/App/components/ui/ToastProvider.js
@@ -1,0 +1,20 @@
+// components/ui/ToastProvider.js
+'use client'
+
+import { Toaster } from 'react-hot-toast'
+
+export default function ToastProvider() {
+  return (
+    <Toaster
+      position="top-center"
+      reverseOrder={false}
+      toastOptions={{
+        style: {
+          borderRadius: '8px',
+          background: '#333',
+          color: '#fff',
+        },
+      }}
+    />
+  )
+}
