+++ Sprint_4/App/package.json
@@ -22,6 +22,8 @@
     ]
   },
   "dependencies": {
+    "@headlessui/react": "^2.2.0",
+    "@heroicons/react": "^1.0.6",
     "@radix-ui/react-avatar": "^1.1.1",
     "@radix-ui/react-dialog": "^1.1.2",
     "@radix-ui/react-label": "^2.1.0",
@@ -50,6 +52,7 @@
     "react-big-calendar": "^1.15.0",
     "react-day-picker": "^8.10.1",
     "react-dom": "^18.3.1",
+    "react-hot-toast": "^2.4.1",
     "react-icons": "^5.3.0",
     "react-modal": "^3.16.1",
     "react-toastify": "^10.0.6",
